requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.40.0
peft>=0.10.0
bitsandbytes>=0.43.0
accelerate>=0.29.0
trl>=0.8.0

# RAG / Vector search
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0

# Data / PDF
datasets>=2.19.0
pypdf>=4.0.0
numpy<2.0
requests==2.32.4

# Dev
jupyter>=1.0.0
ipykernel>=6.0.0
